Refactorings: 1. Moved run_create_vm.py one module up 2. Moved BUCKET_PREFIX constant into asset_id.py 3. Moved vm_slc_bucket.py into it's own submodule
