+++ exasol_script_languages_developer_sandbox/lib/vm_bucket/vm_slc_bucket.py
@@ -8,9 +8,6 @@ BUCKET_NAME = "VMSLCBucket"
 ROLE_NAME = "VMImportRole"
 
 
-BUCKET_PREFIX = "slc_developer_sandbox"
-
-
 def run_setup_vm_bucket(aws_access: AwsAccess) -> None:
     yml = render_template("vm_bucket_cloudformation.jinja.yaml", bucket_name=BUCKET_NAME, role_name=ROLE_NAME)
     aws_access.upload_cloudformation_stack(yml, STACK_NAME)
